AE-99 refactor: add apps API service
